flesh out more code and add vector tests

+++ game/rect.go
@@ -1,10 +1,10 @@
 package game
 
 type Rect struct {
-	X int
-	Y int
-	W int
-	H int
+	X float64
+	Y float64
+	W float64
+	H float64
 }
 
 func (r *Rect) Contains(p Point) bool {
